requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
opencv-python>=4.8.0
torch>=2.6.0
torchvision>=0.21.0
numpy>=1.26.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
aiofiles>=23.0.0